Вынесен общего кода в базовый презентер BaseTraderMvpPresenter из: TraderMeMainPresenter; TraderMainPresenter; SubscriberMainPresenter. Этап 2.
