+++ app/src/main/java/ru/fabulus/fabulustrade/mvp/presenter/base/BaseTraderMvpPresenter.kt
@@ -1,6 +1,7 @@
 package ru.fabulus.fabulustrade.mvp.presenter.base
 
 import android.graphics.Bitmap
+import android.graphics.Color
 import android.util.Log
 import com.github.terrakok.cicerone.Router
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
@@ -8,10 +9,13 @@ import moxy.MvpPresenter
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
 import okhttp3.RequestBody.Companion.toRequestBody
+import ru.fabulus.fabulustrade.R
 import ru.fabulus.fabulustrade.mvp.model.entity.Profile
+import ru.fabulus.fabulustrade.mvp.model.entity.TraderStatistic
 import ru.fabulus.fabulustrade.mvp.model.repo.ApiRepo
 import ru.fabulus.fabulustrade.mvp.model.resource.ResourceProvider
 import ru.fabulus.fabulustrade.mvp.view.base.BaseTraderView
+import ru.fabulus.fabulustrade.util.formatDigitWithDef
 import java.io.ByteArrayOutputStream
 import java.text.SimpleDateFormat
 import java.util.*
@@ -20,9 +24,9 @@ import javax.inject.Inject
 abstract class BaseTraderMvpPresenter<View : BaseTraderView> : MvpPresenter<View>() {
 
     companion object {
-        private const val LOG = "VVV"
         private const val AVATAR_NAME = "avatar"
         private const val DATE_TIME_PATTERN = "dd_MM_yyyy_hh_mm_ss"
+        private const val AVATAR_FILE_EXTENSION = "jpeg"
     }
 
     @Inject
@@ -38,7 +42,6 @@ abstract class BaseTraderMvpPresenter<View : BaseTraderView> : MvpPresenter<View
     lateinit var resourceProvider: ResourceProvider
 
     fun setImage(image: Bitmap) {
-        Log.d(LOG, "BaseTraderMvpPresenter. setImage()")
         val stream = ByteArrayOutputStream()
         image.compress(Bitmap.CompressFormat.JPEG, 80, stream)
         val byteArray = stream.toByteArray()
@@ -49,14 +52,13 @@ abstract class BaseTraderMvpPresenter<View : BaseTraderView> : MvpPresenter<View
                     DATE_TIME_PATTERN,
                     Locale.getDefault()
                 ).format(Date())
-            }.jpeg",
+            }.$AVATAR_FILE_EXTENSION",
             byteArray.toRequestBody("$AVATAR_NAME/*".toMediaTypeOrNull(), 0, byteArray.size)
         )
         changeAvatar(body)
     }
 
     private fun changeAvatar(body: MultipartBody.Part) {
-        Log.d(LOG, "BaseTraderMvpPresenter. changeAvatar()")
         profile.token?.let { token ->
             apiRepo
                 .changeAvatar(token, body)
@@ -67,17 +69,16 @@ abstract class BaseTraderMvpPresenter<View : BaseTraderView> : MvpPresenter<View
                         .observeOn(AndroidSchedulers.mainThread())
                         .subscribe({
                             viewState.setAvatar(it.avatar)
-                        }, {
-                            // Ошибка не обрабатывается
+                        }, { throwable ->
+                            Log.e(BaseTraderMvpPresenter::class.java.name, throwable.message ?: "")
                         })
-                }, {
-                    // Ошибка не обрабатывается
+                }, { throwable ->
+                    Log.e(BaseTraderMvpPresenter::class.java.name, throwable.message ?: "")
                 })
         }
     }
 
     fun deleteAvatar() {
-        Log.d(LOG, "BaseTraderMvpPresenter. deleteAvatar()")
         profile.token?.let { token ->
             apiRepo
                 .deleteAvatar(token)
@@ -88,17 +89,33 @@ abstract class BaseTraderMvpPresenter<View : BaseTraderView> : MvpPresenter<View
                         .observeOn(AndroidSchedulers.mainThread())
                         .subscribe({
                             viewState.setAvatar(it.avatar)
-                        }, {
-                            // Ошибка не обрабатывается
+                        }, { throwable ->
+                            Log.e(BaseTraderMvpPresenter::class.java.name, throwable.message ?: "")
                         })
-                }, {
-                    // Ошибка не обрабатывается
+                }, { throwable ->
+                    Log.e(BaseTraderMvpPresenter::class.java.name, throwable.message ?: "")
                 })
         }
     }
 
+    protected fun getProfitAndColor(traderStatistic: TraderStatistic): Pair<Int, String> {
+        var tmpColor = resourceProvider.getColor(R.color.colorDarkGray)
+        var tmpProfit = resourceProvider.getStringResource(R.string.empty_profit_result)
+
+        traderStatistic.colorIncrDecrDepo365?.let { colorItem ->
+            tmpProfit = resourceProvider.formatDigitWithDef(
+                R.string.incr_decr_depo_365,
+                colorItem.value
+            )
+
+            colorItem.color?.let { color ->
+                tmpColor = Color.parseColor(color)
+            }
+        }
+        return Pair(tmpColor, tmpProfit)
+    }
+
     fun backClicked(): Boolean {
-        Log.d(LOG, "BaseTraderMvpPresenter. backClicked()")
         router.exit()
         return true
     }
