+++ app/src/main/java/ru/fabulus/fabulustrade/mvp/presenter/trader/TraderMainPresenter.kt
@@ -1,7 +1,5 @@
 package ru.fabulus.fabulustrade.mvp.presenter.trader
 
-import android.graphics.Color
-import android.util.Log
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import ru.fabulus.fabulustrade.R
 import ru.fabulus.fabulustrade.mvp.model.entity.Trader
@@ -9,20 +7,17 @@ import ru.fabulus.fabulustrade.mvp.model.entity.TraderStatistic
 import ru.fabulus.fabulustrade.mvp.presenter.base.BaseTraderMvpPresenter
 import ru.fabulus.fabulustrade.mvp.view.trader.TraderMainView
 import ru.fabulus.fabulustrade.navigation.Screens
-import ru.fabulus.fabulustrade.util.formatDigitWithDef
 import java.net.ProtocolException
 
 class TraderMainPresenter(val trader: Trader) : BaseTraderMvpPresenter<TraderMainView>() {
     companion object {
         private const val OBSERVER = 1
         private const val TRADER = 2
-        private const val LOG = "VVV"
     }
 
     private var isObserveActive: Boolean = false
 
     override fun onFirstViewAttach() {
-        Log.d(LOG, "TraderMainPresenter. onFirstViewAttach()")
         super.onFirstViewAttach()
         loadTraderStatistic()
         setVisibility(false)
@@ -33,26 +28,11 @@ class TraderMainPresenter(val trader: Trader) : BaseTraderMvpPresenter<TraderMai
     }
 
     private fun loadTraderStatistic() {
-        Log.d(LOG, "TraderMainPresenter. loadTraderStatistic()")
         apiRepo
             .getTraderStatistic(trader.id)
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe({ traderStatistic: TraderStatistic ->
-
-                var tmpColor = resourceProvider.getColor(R.color.colorDarkGray)
-                var tmpProfit = resourceProvider.getStringResource(R.string.empty_profit_result)
-
-                traderStatistic.colorIncrDecrDepo365?.let { colorItem ->
-                    tmpProfit = resourceProvider.formatDigitWithDef(
-                        R.string.incr_decr_depo_365,
-                        colorItem.value
-                    )
-
-                    colorItem.color?.let { color ->
-                        tmpColor = Color.parseColor(color)
-                    }
-                }
-
+                val (tmpColor, tmpProfit) = getProfitAndColor(traderStatistic)
                 viewState.setProfit(tmpProfit, tmpColor)
                 viewState.initVP(traderStatistic, trader)
             }, { error ->
@@ -61,7 +41,6 @@ class TraderMainPresenter(val trader: Trader) : BaseTraderMvpPresenter<TraderMai
     }
 
     fun observeBtnClicked(subscribe: Boolean) {
-        Log.d(LOG, "TraderMainPresenter. observeBtnClicked()")
         viewState.setObserveChecked(subscribe)
         when {
             profile.user == null -> {
@@ -91,7 +70,6 @@ class TraderMainPresenter(val trader: Trader) : BaseTraderMvpPresenter<TraderMai
     }
 
     fun subscribeToTraderBtnClicked() {
-        Log.d(LOG, "TraderMainPresenter. subscribeToTraderBtnClicked()")
         when {
             profile.user == null -> {
                 router.navigateTo(Screens.signInScreen(false))
@@ -127,7 +105,6 @@ class TraderMainPresenter(val trader: Trader) : BaseTraderMvpPresenter<TraderMai
     }
 
     private fun checkSubscription() {
-        Log.d(LOG, "TraderMainPresenter. checkSubscription()")
         if (profile.user != null) {
             setVisibility(false)
             apiRepo
@@ -163,7 +140,6 @@ class TraderMainPresenter(val trader: Trader) : BaseTraderMvpPresenter<TraderMai
     }
 
     private fun setVisibility(result: Boolean) {
-        Log.d(LOG, "TraderMainPresenter. setVisibility()")
         isObserveActive = result
         viewState.setSubscribeBtnActive(result)
         viewState.setObserveVisibility(result)
